CLI: Move parseOptionsAndRunCLI() to run-cli.ts
